Move JsonDict type into _types package
