+++ tamr_client/_types/__init__.py
@@ -1,0 +1,1 @@
+from .json import JsonDict
