+++ tamr_client/dataset/record.py
@@ -8,9 +8,9 @@ import json
 from typing import cast, Dict, IO, Iterable, Iterator, Optional
 
 from tamr_client import response
+from tamr_client._types import JsonDict
 from tamr_client.dataset.dataset import AnyDataset, Dataset
 from tamr_client.session import Session
-from tamr_client.types import JsonDict
 
 
 class PrimaryKeyNotFound(Exception):
